Скрипт js сделан модульным, разбит на блоки
