+++ source/index.html
@@ -218,7 +218,7 @@
     </div>
   </footer>
 
-  <script src="./js/script.js"></script>
+  <script type="module" src="./js/script.js"></script>
 </body>
 
 </html>
