+++ source/sass/block/main-nav.scss
@@ -1,10 +1,44 @@
 .main-nav {
 
   display: none;
+  opacity: 0;
+
+  animation: fadeInFromNone 0.5s ease-in-out;
+
+  @keyframes fadeInFromNone {
+    0% {
+      opacity: 1;
+    }
+
+    100% {
+      opacity: 0;
+    }
+  }
+
 
   &--show {
     display: block;
     position: relativ;
+    opacity: 1;
+
+    animation: fadeInFromNone 0.5s ease-in-out;
+  }
+
+  @keyframes fadeInFromNone {
+    0% {
+      display: none;
+      opacity: 0;
+    }
+
+    1% {
+      display: block;
+      opacity: 0;
+    }
+
+    100% {
+      display: block;
+      opacity: 1;
+    }
   }
 
   &__container {
