+++ source/sass/block/slider.scss
@@ -1,6 +1,6 @@
 .slider {
   position: relative;
-  overflow: hidden;
+  // overflow: hidden;
   width: 100%;
   background-color: $basic-grey;
